+++ test/mock/create-diff-tap.cjs
