+++ test/mock/create-empty-tap.cjs
