+++ test/mock/create-mixed-tap.cjs
