+++ test/mock/create-passing-tap.cjs
