+++ test/mock/create-throws-tap.cjs
