+++ test/mock/missing-assertions-passing.txt
@@ -1,0 +1,31 @@
+TAP version 13
+# basic arithmetic
+ok 1 should be strictly equal
+ok 2 should be strictly equal
+# deep equality
+ok 3 should be deeply equivalent
+# comparing booleans
+ok 5 should be truthy
+# negatives
+ok 6 should not be strictly equal
+ok 7 should not be deeply equivalent
+# empty map
+# map with elements
+ok 9 (unnamed assert)
+ok 10 (unnamed assert)
+# more info
+ok 11 basic arithmetic still works
+# asynchronous results
+ok 13 should be strictly equal
+ok 14 should be strictly equal
+# nested
+# (anonymous)
+ok 15 should be strictly equal
+# (anonymous)
+ok 16 (unnamed assert)
+
+1..16
+# tests 16
+# pass  16
+
+# ok
